#102 Creates InvolvedCompanyGameCreateFacade
